--- a/.gitignore
+++ b/.gitignore
@@ -3,3 +3,55 @@ test.py
 trained_model.pth
 dataset.txt
 checkpoints/*
+# Byte-compiled / optimized / DLL files
+__pycache__/
+*.py[cod]
+*$py.class
+
+# Distribution / packaging
+dist/
+build/
+*.egg-info/
+
+# Virtual environments
+venv/
+env/
+.env/
+.venv/
+ENV/
+
+# IDE settings
+.idea/
+.vscode/
+*.swp
+*.swo
+
+# Python testing
+.pytest_cache/
+.coverage
+htmlcov/
+
+# Jupyter Notebook
+.ipynb_checkpoints
+
+# Python virtual environment additional files
+.Python
+pip-log.txt
+pip-delete-this-directory.txt
+*.pyc
+*.pyo
+*.pyd
+.Python
+*.so
+.python-version
+.env.fish
+fish/
+__pycache__/
+*.py[cod]
+
+# Fish shell
+fish_history
+fishd.*
+.config/fish/
+.local/share/fish/
+