--- a/requirements.txt
+++ b/requirements.txt
@@ -1,3 +1,3 @@
 torch
 cryptography
-#transformers for the demo
+transformers